Renamed provenance module and added interface imports to __init__
