+++ elephant/buffalo/decorator.py
