+++ elephant/buffalo/examples/run_isi_histogram.py
@@ -4,8 +4,6 @@ import os
 import numpy as np
 
 
-# matplotlib.use('qt5agg')
-
 import matplotlib.pyplot as plt
 import quantities as pq
 
@@ -14,17 +12,17 @@ from reachgraspio import ReachGraspIO
 from elephant.statistics import isi, mean_firing_rate, fanofactor
 from elephant.spike_train_generation import homogeneous_poisson_process
 
-from elephant.buffalo import provenance
+from elephant.buffalo import decorator
 from elephant.buffalo.examples.utils.files import get_file_name
 
 import warnings
 
 plt.Figure.savefig = \
-    provenance.Provenance(inputs=['self'],
-                          file_output=['fname'])(plt.Figure.savefig)
+    decorator.Provenance(inputs=['self'],
+                         file_output=['fname'])(plt.Figure.savefig)
 
 
-@provenance.Provenance(inputs=["isi_times"])
+@decorator.Provenance(inputs=["isi_times"])
 def plot_isi_histograms(grid, *isi_times, bin_size=2*pq.ms, max_time=500*pq.ms,
                         titles=None):
     """
@@ -114,7 +112,7 @@ def plot_isi_histograms(grid, *isi_times, bin_size=2*pq.ms, max_time=500*pq.ms,
     return fig, axes
 
 
-@provenance.Provenance(inputs=[], file_input=['session_filename'])
+@decorator.Provenance(inputs=[], file_input=['session_filename'])
 def load_data(session_filename):
     """
     Loads Reach2Grasp data using the custom BlackRockIO object ReachGraspIO.
@@ -137,14 +135,14 @@ def load_data(session_filename):
                            verbose=False)
 
     block = session.read_block(load_waveforms=False, nsx_to_load=None,
-                               load_events=True, lazy=False, channels='all',
+                               load_events=True, lazy=False, channels=[10],
                                units='all')
 
     return block
 
 
 def main(session_filename):
-    provenance.activate()
+    decorator.activate()
 
     # Load the data
     block = load_data(session_filename)
@@ -174,8 +172,8 @@ def main(session_filename):
     figure.savefig("isi.png")
 
     # provenance.print_history()
-    provenance.save_graph(get_file_name(__file__, extension=".html"),
-                          show=True)
+    decorator.save_graph(get_file_name(__file__, extension=".html"),
+                         show=True)
 
 
 if __name__ == "__main__":
